+++ packages/service-library/src/servicelib/rabbitmq/__init__.py
@@ -1,9 +1,9 @@
-from .constants import BIND_TO_ALL_TOPICS
-from .errors import RemoteMethodNotRegisteredError, RPCNotInitializedError
-from .models import RPCMethodName, RPCNamespace
-from .rabbitmq import RabbitMQClient
-from .rpc_router import RPCRouter
-from .utils import rpc_register_entries
+from ._client import RabbitMQClient
+from ._constants import BIND_TO_ALL_TOPICS
+from ._errors import RemoteMethodNotRegisteredError, RPCNotInitializedError
+from ._models import RPCMethodName, RPCNamespace
+from ._rpc_router import RPCRouter
+from ._utils import rpc_register_entries
 
 __all__: tuple[str, ...] = (
     "BIND_TO_ALL_TOPICS",
