+++ packages/service-library/src/servicelib/rabbitmq/_client.py
@@ -1,49 +1,26 @@
 import asyncio
-import functools
 import logging
-from collections.abc import Callable
 from dataclasses import dataclass, field
-from typing import Any, Final
+from typing import Final
 
 import aio_pika
-import aiormq
-from aio_pika.patterns import RPC
-from pydantic import PositiveInt
-from settings_library.rabbit import RabbitSettings
 
-from ..logging_utils import log_catch, log_context
-from .errors import RemoteMethodNotRegisteredError, RPCNotInitializedError
-from .models import (
-    MessageHandler,
-    RabbitMessage,
-    RPCMethodName,
-    RPCNamespace,
-    RPCNamespacedMethodName,
-)
-from .rpc_router import RPCRouter
-from .utils import declare_queue, get_rabbitmq_client_unique_name
+from ..logging_utils import log_context
+from ._client_base import RabbitMQClientBase
+from ._models import MessageHandler, RabbitMessage
+from ._utils import declare_queue, get_rabbitmq_client_unique_name
 
 _logger = logging.getLogger(__name__)
 
 
-_DEFAULT_RABBITMQ_SERVER_HEARTBEAT_S: Final[int] = 60
 _DEFAULT_PREFETCH_VALUE: Final[int] = 10
 _DEFAULT_RABBITMQ_EXECUTION_TIMEOUT_S: Final[int] = 5
 
 
 @dataclass
-class RabbitMQClient:
-    client_name: str
-    settings: RabbitSettings
-    heartbeat: int = _DEFAULT_RABBITMQ_SERVER_HEARTBEAT_S
+class RabbitMQClient(RabbitMQClientBase):
     _connection_pool: aio_pika.pool.Pool | None = field(init=False, default=None)
     _channel_pool: aio_pika.pool.Pool | None = field(init=False, default=None)
-
-    _rpc_connection: aio_pika.abc.AbstractConnection | None = None
-    _rpc_channel: aio_pika.abc.AbstractChannel | None = None
-    _rpc: RPC | None = None
-
-    _healthy_state: bool = True
 
     def __post_init__(self) -> None:
         # recommendations are 1 connection per process
@@ -52,42 +29,6 @@ class RabbitMQClient:
         )
         # channels are not thread safe, what about python?
         self._channel_pool = aio_pika.pool.Pool(self._get_channel, max_size=10)
-
-    def _connection_close_callback(
-        self,
-        sender: Any,  # pylint: disable=unused-argument
-        exc: BaseException | None,
-    ) -> None:
-        if exc:
-            if isinstance(exc, asyncio.CancelledError):
-                _logger.info("Rabbit connection cancelled")
-            elif isinstance(exc, aiormq.exceptions.ConnectionClosed):
-                _logger.info("Rabbit connection closed: %s", exc)
-            else:
-                _logger.error(
-                    "Rabbit connection closed with exception from %s:%s",
-                    type(exc),
-                    exc,
-                )
-                self._healthy_state = False
-
-    def _channel_close_callback(
-        self,
-        sender: Any,  # pylint: disable=unused-argument  # noqa: ARG002
-        exc: BaseException | None,
-    ) -> None:
-        if exc:
-            if isinstance(exc, asyncio.CancelledError):
-                _logger.info("Rabbit channel cancelled")
-            elif isinstance(exc, aiormq.exceptions.ChannelClosed):
-                _logger.info("Rabbit channel closed")
-            else:
-                _logger.error(
-                    "Rabbit channel closed with exception from %s:%s",
-                    type(exc),
-                    exc,
-                )
-                self._healthy_state = False
 
     async def _get_connection(
         self, rabbit_broker: str, connection_name: str
@@ -104,22 +45,6 @@ class RabbitMQClient:
         connection.close_callbacks.add(self._connection_close_callback)
         return connection
 
-    @property
-    def healthy(self) -> bool:
-        return self._healthy_state
-
-    async def rpc_initialize(self) -> None:
-        self._rpc_connection = await aio_pika.connect_robust(
-            self.settings.dsn,
-            client_properties={
-                "connection_name": f"{get_rabbitmq_client_unique_name(self.client_name)}.rpc"
-            },
-        )
-        self._rpc_channel = await self._rpc_connection.channel()
-
-        self._rpc = RPC(self._rpc_channel)
-        await self._rpc.initialize()
-
     async def close(self) -> None:
         with log_context(
             _logger,
@@ -131,27 +56,12 @@ class RabbitMQClient:
             assert self._connection_pool  # nosec
             await self._connection_pool.close()
 
-            # rpc is not always initialized
-            if self._rpc is not None:
-                await self._rpc.close()
-            if self._rpc_channel is not None:
-                await self._rpc_channel.close()
-            if self._rpc_connection is not None:
-                await self._rpc_connection.close()
-
     async def _get_channel(self) -> aio_pika.abc.AbstractChannel:
         assert self._connection_pool  # nosec
         async with self._connection_pool.acquire() as connection:
             channel = await connection.channel()
             channel.close_callbacks.add(self._channel_close_callback)
             return channel
-
-    async def ping(self) -> bool:
-        with log_catch(_logger, reraise=False):
-            async with await aio_pika.connect(self.settings.dsn, timeout=1):
-                ...
-            return True
-        return False
 
     async def subscribe(
         self,
@@ -302,88 +212,3 @@ class RabbitMQClient:
                 aio_pika.Message(message.body()),
                 routing_key=message.routing_key() or "",
             )
-
-    async def rpc_request(
-        self,
-        namespace: RPCNamespace,
-        method_name: RPCMethodName,
-        *,
-        timeout_s: PositiveInt | None = 5,
-        **kwargs,
-    ) -> Any:
-        """
-        Call a remote registered `handler` by providing it's `namespace`, `method_name`
-        and `kwargs` containing the key value arguments expected by the remote `handler`.
-
-        :raises asyncio.TimeoutError: when message expired
-        :raises CancelledError: when called :func:`RPC.cancel`
-        :raises RuntimeError: internal error
-        :raises RemoteMethodNotRegisteredError: when no handler was registered to the
-            `namespaced_method_name`
-        """
-
-        if not self._rpc:
-            raise RPCNotInitializedError
-
-        namespaced_method_name = RPCNamespacedMethodName.from_namespace_and_method(
-            namespace, method_name
-        )
-        try:
-            queue_expiration_timeout = timeout_s
-            awaitable = self._rpc.call(
-                namespaced_method_name,
-                expiration=queue_expiration_timeout,
-                kwargs=kwargs,
-            )
-            return await asyncio.wait_for(awaitable, timeout=timeout_s)
-        except aio_pika.MessageProcessError as e:
-            if e.args[0] == "Message has been returned":
-                raise RemoteMethodNotRegisteredError(
-                    method_name=namespaced_method_name, incoming_message=e.args[1]
-                ) from e
-            raise
-
-    async def rpc_register_handler(
-        self,
-        namespace: RPCNamespace,
-        method_name: RPCMethodName,
-        handler: Callable[..., Any],
-    ) -> None:
-        """
-        Bind a local `handler` to a `namespace` and `method_name`.
-        The handler can be remotely called by providing the `namespace` and `method_name`
-
-        NOTE: method_name could be computed from the handler, but by design, it is
-        left to the caller to do so.
-        """
-
-        if self._rpc is None:
-            raise RPCNotInitializedError
-
-        await self._rpc.register(
-            RPCNamespacedMethodName.from_namespace_and_method(namespace, method_name),
-            handler,
-            auto_delete=True,
-        )
-
-    async def rpc_register_router(
-        self,
-        router: RPCRouter,
-        namespace: RPCNamespace,
-        *handler_args,
-        **handler_kwargs,
-    ) -> None:
-        for rpc_method_name, handler in router.routes.items():
-            await self.rpc_register_handler(
-                namespace,
-                rpc_method_name,
-                functools.partial(handler, *handler_args, **handler_kwargs),
-            )
-
-    async def rpc_unregister_handler(self, handler: Callable[..., Any]) -> None:
-        """Unbind a locally added `handler`"""
-
-        if self._rpc is None:
-            raise RPCNotInitializedError
-
-        await self._rpc.unregister(handler)
