+++ packages/service-library/src/servicelib/rabbitmq/_constants.py
