+++ packages/service-library/src/servicelib/rabbitmq/_utils.py
@@ -11,7 +11,7 @@ from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
 
 from ..logging_utils import log_context
-from .models import RPCMethodName, RPCNamespace
+from ._models import RPCMethodName, RPCNamespace
 
 _logger = logging.getLogger(__file__)
